refactor: flatten folder and change path
