+++ src/modules/CourseSearch/components/CourseCard/components/CardDescription/index.tsx
