+++ src/modules/CourseSearch/components/CourseCard/components/CardHiddenDescription/index.tsx
