+++ src/modules/CourseSearch/components/CourseCard/index.tsx
@@ -1,12 +1,12 @@
 import { Paper } from '@material-ui/core'
 import { forwardRef, memo } from 'react'
 
-import { CardAction } from '@/modules/CourseSearch/components/CourseCard/components/CardAction'
-import { CardBody } from '@/modules/CourseSearch/components/CourseCard/components/CardBody'
-import { CardHead } from '@/modules/CourseSearch/components/CourseCard/components/CardHead'
-import { CourseCardContext } from '@/modules/CourseSearch/components/CourseCard/context'
-import { useCourseCardContext } from '@/modules/CourseSearch/components/CourseCard/hooks/useCourseCard'
-import { CourseCardProps } from '@/modules/CourseSearch/components/CourseCard/types'
+import { CardAction } from './components/CardAction'
+import { CardBody } from './components/CardBody'
+import { CardHead } from './components/CardHead'
+import { CourseCardContext } from './context'
+import { useCourseCardContext } from './hooks/useCourseCard'
+import { CourseCardProps } from './types'
 
 export const CourseCard = memo(
   forwardRef<HTMLDivElement, CourseCardProps>(function CourseCard(props, ref) {
